pyproject.toml
[project]
name = "autohwformer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "easydict>=1.13",
    "gdown>=5.2.0",
    "onnxscript>=0.2.2",
    "opencv-python>=4.11.0.86",
    "psutil>=7.0.0",
    "python-hostlist>=2.2.1",
    "scikit-image>=0.25.2",
    "timm>=1.0.15",
    "torch==2.5.0",
    "tqdm>=4.67.1",
    "uv>=0.6.6",
    "wandb>=0.19.6",
    "zigzag-dse>=3.8.1",
]

[project.optional-dependencies]
notebook = [
    "ipykernel>=6.29.5",
]
profile = [
    "snakeviz>=2.2.2",
]

line-length = 120
indent-width = 4

[lint]
select = [
    # mccabe (C90), commas (COM), comprehensions (C4)
    "C",
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # isort
    "I",
    # pylint
    "PL",
    # pycodestyle warnings
    "W",
]

# Ignore import violations in all `__init__.py` files.
[lint.per-file-ignores]
"__init__.py" = ["E402", "F401", "F403", "F811"]

[lint.isort]
lines-after-imports = 2

[format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"